feat(core): Expr/NamedTuple/Pair/Function/Sparse codecs; JLD2 is a dependency (Closes #759)
