+++ Project.toml
@@ -12,6 +12,7 @@ Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
@@ -32,13 +33,11 @@ TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [weakdeps]
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 PATHSolver = "f5f7c340-0bb3-5c69-969a-41884d311d1b"
 XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
 ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
 
 [extensions]
-MacroEconometricModelsJLD2Ext = ["JLD2"]
 MacroEconometricModelsPATHExt = ["PATHSolver"]
 MacroEconometricModelsXLSXExt = ["XLSX"]
 MacroEconometricModelsZipFileExt = ["ZipFile"]
@@ -82,11 +81,10 @@ julia = "1.10"
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 PATHSolver = "f5f7c340-0bb3-5c69-969a-41884d311d1b"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
 ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
 
 [targets]
-test = ["Aqua", "Test", "Documenter", "JLD2", "PATHSolver", "XLSX", "ZipFile"]
+test = ["Aqua", "Test", "Documenter", "PATHSolver", "XLSX", "ZipFile"]
