--- src/core/serialization.jl
@@ -1,509 +1,0 @@
-# MacroEconometricModels.jl
-# Copyright (C) 2025-2026 Wookyung Chung <chung@friedman.jp>
-#
-# This file is part of MacroEconometricModels.jl.
-# Licensed under GPL-3.0-or-later. See LICENSE for details.
-
-# =============================================================================
-# Versioned result serialization (T248 / #347, extended to full coverage #505)
-# =============================================================================
-# `save_model(result, path)` / `load_model(path)` persist a fitted model — or a
-# data container — to disk in a self-describing, version-tagged container so
-# files survive a package upgrade. Bare `Base.serialize` is deliberately NOT the
-# on-disk format — it breaks across Julia versions and struct-layout changes.
-#
-# Design:
-#   1. `_to_serializable(m)` reduces a result to a `Dict{String,Any}` of PLAIN
-#      values only — numbers, strings, symbols, bools, arrays, `nothing`, and
-#      nested `Dict`s. No custom struct survives into the payload: the
-#      reproducibility manifest, the LP covariance estimator, and every *nested*
-#      MacroEconometricModels struct (a `VECMModel`'s `JohansenResult`, an
-#      `IOData`'s `IOMetaData`, a `StructuralDFM`'s wrapped `VARModel`, …) are
-#      flattened recursively to tagged dicts. Storing only primitives is exactly
-#      what makes the format robust across versions.
-#   2. `_build_container(m)` wraps that payload with a metadata header: the
-#      `format_version`, the package + Julia versions, a timestamp, the result
-#      type name, and (when present) the reproducibility manifest.
-#   3. The actual disk read/write is a JLD2 weak-dependency backend
-#      (`_write_model_container` / `_read_model_container`, overridden in
-#      ext/MacroEconometricModelsJLD2Ext.jl). Without JLD2 loaded the stub raises
-#      an informative "]add JLD2" error. ALL model↔dict logic and version
-#      validation stay here in src, so they are testable without the backend.
-#   4. `load_model` validates the `format_version` and type tag and raises a
-#      typed `SerializationError` naming the expected-vs-found version on a
-#      mismatch, rather than returning a corrupted object.
-#
-# Reduction is generic — `_capture_fields` walks the public fields and
-# `_ser_field` flattens each — so adding a type is largely a matter of listing it
-# in `_SERIALIZABLE_TYPES`. Reconstruction is a generic positional-constructor
-# call (`_generic_construct`), overridden explicitly only where a type takes a
-# keyword-defaulted field (e.g. `BVARPosterior`'s `manifest`).
-#
-# NOT yet covered: the DSGE / HA / OLG / CT solution family
-# (`DSGESolution`, `BayesianDSGE`, `HADSGESolution`, …). Those transitively hold
-# a `ModelSpec`, whose `residual_fns::Vector{Function}` / `ss_fn` are compiled
-# closures generated by the `@dsge` macro — functions do not round-trip through a
-# version-stable payload, and rebuilding them requires re-parsing the stored
-# equations at load time. That is a distinct effort tracked as follow-up work.
-
-"""
-    SERIALIZATION_FORMAT_VERSION
-
-On-disk schema version written into every model file by [`save_model`](@ref) and
-checked by [`load_model`](@ref). Bumped only on a breaking change to the payload
-layout; a file whose version this build does not recognize is rejected with a
-[`SerializationError`](@ref) rather than mis-read.
-"""
-const SERIALIZATION_FORMAT_VERSION = 1
-
-# Result / data types with round-trip support. Maps the stored type tag → the
-# concrete type, so `load_model` dispatches `_from_serializable` on the recorded
-# name. Only TOP-LEVEL `save_model` targets are listed here; nested structs are
-# resolved by name at load time (see `_resolve_ser_type`) and so need no entry.
-const _SERIALIZABLE_TYPES = Dict{String,Type}(
-    # ── core VAR family (original T248 set) ──────────────────────────────────
-    "VARModel"                      => VARModel,
-    "BVARPosterior"                 => BVARPosterior,
-    "RegModel"                      => RegModel,
-    "LogitModel"                    => LogitModel,
-    "ProbitModel"                   => ProbitModel,
-    "LPModel"                       => LPModel,
-    # ── data containers ──────────────────────────────────────────────────────
-    "TimeSeriesData"                => TimeSeriesData,
-    "PanelData"                     => PanelData,
-    "CrossSectionData"              => CrossSectionData,
-    "IOData"                        => IOData,
-    "IOMetaData"                    => IOMetaData,
-    "ExtractionResult"              => ExtractionResult,
-    "PriceModelResult"              => PriceModelResult,
-    "ImpactResult"                  => ImpactResult,
-    "NetworkStatsResult"            => NetworkStatsResult,
-    "VerticalSpecialization"        => VerticalSpecialization,
-    "ExportDecomposition"           => ExportDecomposition,
-    "RegionalFootprintResult"       => RegionalFootprintResult,
-    "BaqaeeFarhiResult"             => BaqaeeFarhiResult,
-    "ProductionNetwork"             => ProductionNetwork,
-    "SDAResult"                     => SDAResult,
-    "RASResult"                     => RASResult,
-    "BFLocal"                       => BFLocal,
-    "BFElasticities"                => BFElasticities,
-    "BFShockCurve"                  => BFShockCurve,
-    "BFEquilibrium"                 => BFEquilibrium,
-    "BFWedgeDecomp"                 => BFWedgeDecomp,
-    "BFMisallocation"               => BFMisallocation,
-    # ── cointegration / VECM ─────────────────────────────────────────────────
-    "VECMModel"                     => VECMModel,
-    "CointRegModel"                 => CointRegModel,
-    "PanelCointRegModel"            => PanelCointRegModel,
-    # ── panel / PVAR ─────────────────────────────────────────────────────────
-    "PVARModel"                     => PVARModel,
-    "PanelRegModel"                 => PanelRegModel,
-    "PanelIVModel"                  => PanelIVModel,
-    "PanelLogitModel"               => PanelLogitModel,
-    "PanelProbitModel"              => PanelProbitModel,
-    # ── ordered / multinomial choice ─────────────────────────────────────────
-    "OrderedLogitModel"             => OrderedLogitModel,
-    "OrderedProbitModel"            => OrderedProbitModel,
-    "MultinomialLogitModel"         => MultinomialLogitModel,
-    # ── local-projection variants ────────────────────────────────────────────
-    "LPIVModel"                     => LPIVModel,
-    "SmoothLPModel"                 => SmoothLPModel,
-    "StateLPModel"                  => StateLPModel,
-    "PropensityLPModel"             => PropensityLPModel,
-    # ── systems / GMM ────────────────────────────────────────────────────────
-    "SURModel"                      => SURModel,
-    "GMMModel"                      => GMMModel,
-    "SMMModel"                      => SMMModel,
-    # ── volatility ───────────────────────────────────────────────────────────
-    "ARCHModel"                     => ARCHModel,
-    "GARCHModel"                    => GARCHModel,
-    "EGARCHModel"                   => EGARCHModel,
-    "GJRGARCHModel"                 => GJRGARCHModel,
-    "APARCHModel"                   => APARCHModel,
-    "CGARCHModel"                   => CGARCHModel,
-    "FIGARCHModel"                  => FIGARCHModel,
-    "FIEGARCHModel"                 => FIEGARCHModel,
-    "GarchMidasModel"               => GarchMidasModel,
-    "SVModel"                       => SVModel,
-    "MGARCHModel"                   => MGARCHModel,
-    # ── factor / FAVAR ───────────────────────────────────────────────────────
-    "FactorModel"                   => FactorModel,
-    "DynamicFactorModel"            => DynamicFactorModel,
-    "GeneralizedDynamicFactorModel" => GeneralizedDynamicFactorModel,
-    "FAVARModel"                    => FAVARModel,
-    "StructuralDFM"                 => StructuralDFM,
-    # ── ARIMA / ARDL / nonlinear / MIDAS / state space ───────────────────────
-    "ARModel"                       => ARModel,
-    "MAModel"                       => MAModel,
-    "ARMAModel"                     => ARMAModel,
-    "ARIMAModel"                    => ARIMAModel,
-    "ARFIMAModel"                   => ARFIMAModel,
-    "ARDLModel"                     => ARDLModel,
-    "NARDLModel"                    => NARDLModel,
-    "PMGModel"                      => PMGModel,
-    "ThresholdModel"                => ThresholdModel,
-    "MidasModel"                    => MidasModel,
-    "StateSpaceModel"               => StateSpaceModel,
-    # ── SVAR identification results (SID-24 / #753) ──────────────────────────
-    "ProxySVARResult"               => ProxySVARResult,
-    "SVARModel"                     => SVARModel,
-    "MaxShareResult"                => MaxShareResult,
-    "SVECResult"                    => SVECResult,
-    "ICASVARResult"                 => ICASVARResult,
-    "NonGaussianMLResult"           => NonGaussianMLResult,
-    "NonGaussianGMMResult"          => NonGaussianGMMResult,
-    "MarkovSwitchingSVARResult"     => MarkovSwitchingSVARResult,
-    "GARCHSVARResult"               => GARCHSVARResult,
-    "SmoothTransitionSVARResult"    => SmoothTransitionSVARResult,
-    "ExternalVolatilitySVARResult"  => ExternalVolatilitySVARResult,
-    "AriasSVARResult"               => AriasSVARResult,
-    "UhligSVARResult"               => UhligSVARResult,
-    "BayesianSetIdentifiedSVAR"     => BayesianSetIdentifiedSVAR,
-    "SignIdentifiedSet"             => SignIdentifiedSet,
-    "RobustBayesResult"             => RobustBayesResult,
-)
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Field-value flattening (struct → plain values)
-# ─────────────────────────────────────────────────────────────────────────────
-
-# A value whose element type is one of these needs no flattening — JLD2 stores
-# numbers/strings/symbols/bools/chars/enums (and arrays/dicts of them) robustly.
-_is_plain_eltype(::Type{S}) where {S} = S <: Union{Number,AbstractString,Symbol,Char,Bool,Enum}
-
-# Is `x` a MacroEconometricModels struct we should flatten recursively? Types
-# from Base / LinearAlgebra / Distributions (parentmodule ≠ this package) fall
-# through to JLD2's own storage; the manifest and covariance estimators have
-# their own tagged-dict methods below and never reach here.
-function _is_mem_struct(x)
-    T = typeof(x)
-    isstructtype(T) || return false
-    return parentmodule(T) === MacroEconometricModels
-end
-
-# The manifest and LP covariance estimators are flattened by dedicated methods;
-# everything else routes through the generic recursive `_ser_field(x)`.
-_ser_field(x::ReproManifest) = _manifest_to_dict(x)
-_ser_field(x::AbstractCovarianceEstimator) = _cov_to_dict(x)
-
-function _ser_field(x)
-    x === nothing && return nothing
-    x isa Missing && return missing
-    # Compiled closures/functions do not round-trip through a version-stable
-    # payload. The only function field among registered types is the OPTIONAL
-    # `StateSpaceModel.builder` re-fit helper, which reloads as `nothing`.
-    x isa Function && return nothing
-    x isa Enum && return x
-    (x isa Number || x isa AbstractString || x isa Symbol || x isa Char) && return x
-    if x isa AbstractArray
-        _is_plain_eltype(eltype(x)) && return x
-        return map(_ser_field, x)
-    end
-    x isa Tuple && return map(_ser_field, x)
-    if x isa AbstractDict
-        _is_plain_eltype(valtype(x)) && return x
-        return Dict(k => _ser_field(v) for (k, v) in x)
-    end
-    _is_mem_struct(x) && return _struct_to_dict(x)
-    return x
-end
-
-# Generic public-field capture: {fieldname => flattened value} for every field.
-function _capture_fields(m)
-    d = Dict{String,Any}()
-    for f in fieldnames(typeof(m))
-        d[String(f)] = _ser_field(getfield(m, f))
-    end
-    return d
-end
-
-# A nested struct is captured like the top-level payload, plus a `__struct__`
-# tag so `_deser_field` can resolve its concrete type on the way back in.
-function _struct_to_dict(m)
-    d = _capture_fields(m)
-    d["__struct__"] = String(nameof(typeof(m)))
-    return d
-end
-
-_extract_manifest(m) = hasfield(typeof(m), :manifest) ? getfield(m, :manifest) : nothing
-
-# LP covariance estimator ↔ dict. These are small config structs (bandwidth,
-# kernel, prewhiten) — capture the type name + fields, rebuild by name.
-function _cov_to_dict(c::AbstractCovarianceEstimator)
-    d = Dict{String,Any}("__estimator__" => string(nameof(typeof(c))))
-    for f in fieldnames(typeof(c))
-        d[String(f)] = getfield(c, f)
-    end
-    return d
-end
-
-function _cov_from_dict(d::AbstractDict)
-    name = String(d["__estimator__"])
-    if name == "NeweyWestEstimator"
-        return NeweyWestEstimator{Float64}(Int(d["bandwidth"]), _as_symbol(d["kernel"]), Bool(d["prewhiten"]))
-    elseif name == "WhiteEstimator"
-        return WhiteEstimator()
-    elseif name == "DriscollKraayEstimator"
-        return DriscollKraayEstimator{Float64}(Int(d["bandwidth"]), _as_symbol(d["kernel"]))
-    end
-    throw(SerializationError("unknown covariance estimator '$name' in serialized LPModel"))
-end
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Field-value reconstruction (plain values → struct)
-# ─────────────────────────────────────────────────────────────────────────────
-
-_as_symbol(x::Symbol) = x
-_as_symbol(x::AbstractString) = Symbol(x)
-
-# Inverse of `_ser_field`: rebuild manifests / covariance estimators / nested
-# structs from their tagged dicts; recurse into arrays/dicts of them.
-function _deser_field(x)
-    if x isa AbstractDict
-        haskey(x, "__manifest__")  && return _manifest_from_dict(x)
-        haskey(x, "__estimator__") && return _cov_from_dict(x)
-        haskey(x, "__struct__")    && return _deser_struct(x)
-        _is_plain_eltype(valtype(x)) && return x
-        return Dict(k => _deser_field(v) for (k, v) in x)
-    end
-    if x isa AbstractArray
-        _is_plain_eltype(eltype(x)) && return x
-        return map(_deser_field, x)
-    end
-    x isa Tuple && return map(_deser_field, x)
-    return x
-end
-
-# Resolve a nested `__struct__` tag to its concrete type. Nested types need no
-# `_SERIALIZABLE_TYPES` entry — they are looked up by name in the package module.
-function _resolve_ser_type(name::AbstractString)
-    sym = Symbol(name)
-    isdefined(MacroEconometricModels, sym) || throw(SerializationError(
-        "serialized nested type '$name' is not defined in this build of MacroEconometricModels"))
-    T = getfield(MacroEconometricModels, sym)
-    T isa Type || throw(SerializationError("serialized type tag '$name' does not name a type"))
-    return T
-end
-
-_deser_struct(d::AbstractDict) =
-    _from_serializable(_resolve_ser_type(String(d["__struct__"])), d, SERIALIZATION_FORMAT_VERSION)
-
-# Infer a `T<:AbstractFloat` type parameter from the reconstructed field values
-# (the first float scalar/array wins). Returns `nothing` for non-float or
-# non-parametric types, in which case the un-parameterized constructor is used.
-function _infer_float_param(args)
-    for a in args
-        a isa AbstractFloat && return typeof(a)
-        if a isa AbstractArray
-            E = eltype(a)
-            E <: AbstractFloat && return E
-            E <: Complex{<:AbstractFloat} && return real(E)
-        end
-        if a isa Tuple    # e.g. PropensityScoreConfig.trimming::Tuple{T,T}
-            for t in a
-                t isa AbstractFloat && return typeof(t)
-            end
-        end
-    end
-    return nothing
-end
-
-# Construct `T` from positional field values. Parametric types with an inner
-# constructor expose only `T{P}(fields...)`; a struct with an un-parameterized
-# inner constructor (e.g. `VARModel`) exposes only `T(fields...)`; a struct with
-# no inner constructor has both. Try the parameterized form first, fall back.
-function _generic_construct(T, args)
-    if T isa UnionAll
-        P = _infer_float_param(args)
-        if P !== nothing
-            try
-                return T{P}(args...)
-            catch e
-                (e isa MethodError || e isa TypeError) || rethrow()
-            end
-        end
-    end
-    return T(args...)
-end
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Per-type reconstruction
-# ─────────────────────────────────────────────────────────────────────────────
-
-# Generic reconstruction: deserialize each public field and route through the
-# type's real constructor (validation preserved). Forward-safe: a v1 loader
-# knows exactly the v1 field set via `fieldnames(T)`.
-function _from_serializable(::Type{T}, p::AbstractDict, ::Int) where {T}
-    args = Any[]
-    for f in fieldnames(T)
-        key = String(f)
-        # A field added to the struct after the file was written must surface as
-        # the format's typed error, not a raw KeyError (#538).
-        haskey(p, key) || throw(SerializationError(
-            "field '$key' of $(nameof(T)) is missing from the payload — the file " *
-            "was saved by an older package version, before this field existed. " *
-            "Re-create and re-save the object with the current version."))
-        push!(args, _deser_field(p[key]))
-    end
-    return _generic_construct(T, args)
-end
-
-# ── explicit overrides where the constructor is not a plain positional call ──
-
-function _from_serializable(::Type{VARModel}, p::AbstractDict, ::Int)
-    VARModel(p["Y"], p["p"], p["B"], p["U"], p["Sigma"], p["aic"], p["bic"], p["hqic"], p["varnames"])
-end
-
-# `manifest` is a keyword-defaulted constructor argument, not a positional field.
-function _from_serializable(::Type{BVARPosterior}, p::AbstractDict, ::Int)
-    T = eltype(p["B_draws"])
-    mani = _manifest_from_dict(get(p, "manifest", nothing))
-    BVARPosterior{T}(p["B_draws"], p["Sigma_draws"], p["n_draws"], p["p"], p["n"],
-                     p["data"], _as_symbol(p["prior"]), _as_symbol(p["sampler"]),
-                     p["varnames"]; manifest=mani)
-end
-
-function _from_serializable(::Type{RegModel}, p::AbstractDict, ::Int)
-    T = eltype(p["y"])
-    RegModel{T}(p["y"], p["X"], p["beta"], p["vcov_mat"], p["residuals"], p["fitted"],
-                p["ssr"], p["tss"], p["r2"], p["adj_r2"], p["f_stat"], p["f_pval"],
-                p["loglik"], p["aic"], p["bic"], p["varnames"], _as_symbol(p["method"]),
-                _as_symbol(p["cov_type"]), p["weights"], p["Z"], p["endogenous"],
-                p["first_stage_f"], p["sargan_stat"], p["sargan_pval"],
-                p["cragg_donald_f"], p["kleibergen_paap_f"], p["stock_yogo_10pct"])
-end
-
-function _from_serializable(::Type{LogitModel}, p::AbstractDict, ::Int)
-    T = eltype(p["y"])
-    LogitModel{T}(p["y"], p["X"], p["beta"], p["vcov_mat"], p["residuals"], p["fitted"],
-                  p["loglik"], p["loglik_null"], p["pseudo_r2"], p["aic"], p["bic"],
-                  p["varnames"], p["converged"], p["iterations"], _as_symbol(p["cov_type"]))
-end
-
-function _from_serializable(::Type{ProbitModel}, p::AbstractDict, ::Int)
-    T = eltype(p["y"])
-    ProbitModel{T}(p["y"], p["X"], p["beta"], p["vcov_mat"], p["residuals"], p["fitted"],
-                   p["loglik"], p["loglik_null"], p["pseudo_r2"], p["aic"], p["bic"],
-                   p["varnames"], p["converged"], p["iterations"], _as_symbol(p["cov_type"]))
-end
-
-function _from_serializable(::Type{LPModel}, p::AbstractDict, ::Int)
-    cov = _cov_from_dict(p["cov_estimator"])
-    LPModel(p["Y"], p["shock_var"], p["response_vars"], p["horizon"], p["lags"],
-            p["B"], p["residuals"], p["vcov"], p["T_eff"], cov, p["varnames"])
-end
-
-# Nested config struct inside `StateLPModel`: its inner constructor recomputes the
-# derived `F_values` field from (state_var, gamma, threshold, method), so route
-# through that 4-argument constructor rather than the full field list.
-function _from_serializable(::Type{StateTransition}, p::AbstractDict, ::Int)
-    StateTransition(p["state_var"], p["gamma"], p["threshold"], _as_symbol(p["method"]))
-end
-
-# `SVARPattern` only exposes a keyword inner constructor (`long_run`).
-function _from_serializable(::Type{SVARPattern}, p::AbstractDict, ::Int)
-    SVARPattern(p["A"], p["B"]; long_run=_deser_field(p["long_run"]))
-end
-
-# Inner constructor validates `:exact/:over/:under/:set`; coerce a stored string.
-function _from_serializable(::Type{IdentificationStatus}, p::AbstractDict, ::Int)
-    IdentificationStatus(_as_symbol(p["status"]),
-                         Vector{Int}(p["ranks"]),
-                         Vector{Int}(p["orders"]),
-                         Int(p["n_overidentifying"]))
-end
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Container assembly + validation
-# ─────────────────────────────────────────────────────────────────────────────
-
-_to_serializable(m) = _capture_fields(m)
-
-function _build_container(m)
-    tname = string(nameof(typeof(m)))
-    haskey(_SERIALIZABLE_TYPES, tname) || throw(SerializationError(
-        "save_model does not support $(typeof(m)); supported types: " *
-        join(sort(collect(keys(_SERIALIZABLE_TYPES))), ", ")))
-    return Dict{String,Any}(
-        "format_version"  => SERIALIZATION_FORMAT_VERSION,
-        "package_version" => _repro_package_version(),
-        "julia_version"   => string(VERSION),
-        "created"         => _repro_timestamp(),
-        "type"            => tname,
-        "manifest"        => _manifest_to_dict(_extract_manifest(m)),
-        "payload"         => _to_serializable(m),
-    )
-end
-
-function _reconstruct_from_container(container::AbstractDict)
-    ver = get(container, "format_version", nothing)
-    ver isa Integer || throw(SerializationError(
-        "not a MacroEconometricModels model file: missing or non-integer format_version"))
-    ver == SERIALIZATION_FORMAT_VERSION || throw(SerializationError(
-        "unsupported serialization format_version $ver: this build reads version " *
-        "$SERIALIZATION_FORMAT_VERSION. Re-save with the current release, or load with a " *
-        "package version whose SERIALIZATION_FORMAT_VERSION == $ver."))
-    tname = get(container, "type", nothing)
-    tname isa AbstractString || throw(SerializationError("serialized model is missing its type tag"))
-    haskey(_SERIALIZABLE_TYPES, tname) || throw(SerializationError(
-        "serialized type '$tname' is not loadable by this build; supported: " *
-        join(sort(collect(keys(_SERIALIZABLE_TYPES))), ", ")))
-    payload = get(container, "payload", nothing)
-    payload isa AbstractDict || throw(SerializationError("serialized model '$tname' has no payload"))
-    return _from_serializable(_SERIALIZABLE_TYPES[tname], payload, ver)
-end
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Backend stubs (overridden by the JLD2 extension) + public API
-# ─────────────────────────────────────────────────────────────────────────────
-
-# Extension entry points — the real methods live in
-# ext/MacroEconometricModelsJLD2Ext.jl and override these more-specific-arg stubs.
-_write_model_container(path, container) =
-    error("save_model requires the JLD2 package. Run `]add JLD2` and `using JLD2` to enable it.")
-_read_model_container(path) =
-    error("load_model requires the JLD2 package. Run `]add JLD2` and `using JLD2` to enable it.")
-
-"""
-    save_model(model, path) -> path
-
-Persist a fitted `model` — or a data container — to `path` in a versioned,
-self-describing container. Coverage spans every VAR/regression/panel/volatility/
-factor/ARIMA/local-projection/GMM model, SVAR identification results, and the
-data containers (`TimeSeriesData`, `PanelData`, `CrossSectionData`, `IOData`);
-the full set is `MacroEconometricModels._SERIALIZABLE_TYPES`. The file records the
-[`SERIALIZATION_FORMAT_VERSION`](@ref), the package and Julia versions, a
-timestamp, and — for a randomized result — its reproducibility manifest. Only
-public fields are stored; cached factorizations are recomputed on load, and
-compiled equation functions (DSGE) are not yet serializable.
-
-Requires the JLD2 backend: `]add JLD2` then `using JLD2`.
-
-```julia
-using JLD2
-m = estimate_var(Y, 2)
-save_model(m, "model.jld2")
-m2 = load_model("model.jld2")   # identical public fields
-```
-"""
-function save_model(model, path::AbstractString)
-    container = _build_container(model)
-    _write_model_container(String(path), container)
-    return path
-end
-
-"""
-    load_model(path) -> model
-
-Reconstruct a model saved by [`save_model`](@ref). Validates the stored
-`format_version` and type tag; raises a [`SerializationError`](@ref) naming the
-expected-versus-found version on an unrecognized format, rather than returning a
-corrupted object. Requires the JLD2 backend (`using JLD2`).
-"""
-function load_model(path::AbstractString)
-    isfile(path) || throw(SerializationError("no such model file: $path"))
-    container = _read_model_container(String(path))
-    container isa AbstractDict || throw(SerializationError(
-        "file '$path' does not contain a MacroEconometricModels model container"))
-    return _reconstruct_from_container(container)
-end
